Add reanchor-weekly migration script + fix resume make_interval(bigint)
- scripts/reanchor-weekly.ts + npm run db:reanchor: dry-run (read-only, default) / --apply (transactional + audited subscription.reanchored). Aborts if an active weekly sub lacks live coverage. Dry-run verified against dev. - Fix: resume Monday re-anchor passed make_interval(days => bigint); cast idx to int (repo + script). Added a resume anchoring test. 6/6 tests pass, typecheck clean. - Docs: migration plan + roadmap updated.

+++ src/modules/drivers/enrollment.repository.ts
@@ -200,7 +200,7 @@ export class EnrollmentRepository {
           `WITH anchor AS (
              SELECT date_trunc('week', (now() AT TIME ZONE $2)) AS monday_local
            ), ordered AS (
-             SELECT sp.id, row_number() OVER (ORDER BY sp.period_start) - 1 AS idx
+             SELECT sp.id, (row_number() OVER (ORDER BY sp.period_start) - 1)::int AS idx
              FROM subscription_payments sp
              JOIN driver_subscriptions ds ON ds.id = sp.driver_subscription_id
              WHERE ds.driver_id = $1 AND sp.status = 'paid'
